internal-docs-qa/requirements.txt
flask
flask-cors
langchain
faiss-cpu
sentence-transformers
numpy==1.26.4
gunicorn
langchain-community
langchain_huggingface